+++ src/parsers.ts
@@ -1,38 +1,80 @@
 import {
-  CapMatch,
   CapModifier,
-  CoreNotationMatch,
-  DropConstraintsMatch,
-  DropHighMatch,
-  DropLowMatch,
   DropModifier,
   DropOptions,
-  ExplodeMatch,
   ExplodeModifier,
   InternalRollParameters,
-  isCapMatch,
-  isDropConstraintsMatch,
-  isDropHighMatch,
-  isDropLowMatch,
-  isExplodeMatch,
-  isPlusMatch,
-  isReplaceMatch,
-  isRerollMatch,
-  isUniqueMatch,
-  Match,
-  MinusMatch,
   MinusModifier,
   Modifier,
-  PlusMatch,
   PlusModifier,
-  ReplaceMatch,
   ReplaceModifier,
-  RerollMatch,
   RerollModifier,
   RerollOptions,
-  UniqueMatch,
   UniqueModifier
 } from './types'
+
+function isMatcherType<T extends Match>(
+  argument: Match,
+  key: keyof T
+): argument is T {
+  return (argument as T)[key] !== undefined
+}
+
+export type CoreNotationMatch = { coreNotationMatch: string }
+export const isCoreNotationMatch = (match: Match): match is CoreNotationMatch =>
+  isMatcherType<CoreNotationMatch>(match, 'coreNotationMatch')
+
+export type DropHighMatch = { dropHighMatch: string }
+export const isDropHighMatch = (match: Match): match is DropHighMatch =>
+  isMatcherType<DropHighMatch>(match, 'dropHighMatch')
+
+export type DropLowMatch = { dropLowMatch: string }
+export const isDropLowMatch = (match: Match): match is DropLowMatch =>
+  isMatcherType<DropLowMatch>(match, 'dropLowMatch')
+
+export type DropConstraintsMatch = { dropConstraintsMatch: string }
+export const isDropConstraintsMatch = (
+  match: Match
+): match is DropConstraintsMatch =>
+  isMatcherType<DropConstraintsMatch>(match, 'dropConstraintsMatch')
+
+export type ExplodeMatch = { explodeMatch: string }
+export const isExplodeMatch = (match: Match): match is ExplodeMatch =>
+  isMatcherType<ExplodeMatch>(match, 'explodeMatch')
+
+export type UniqueMatch = { uniqueMatch: string }
+export const isUniqueMatch = (match: Match): match is UniqueMatch =>
+  isMatcherType<UniqueMatch>(match, 'uniqueMatch')
+
+export type ReplaceMatch = { replaceMatch: string }
+export const isReplaceMatch = (match: Match): match is ReplaceMatch =>
+  isMatcherType<ReplaceMatch>(match, 'replaceMatch')
+
+export type RerollMatch = { rerollMatch: string }
+export const isRerollMatch = (match: Match): match is RerollMatch =>
+  isMatcherType<RerollMatch>(match, 'rerollMatch')
+
+export type CapMatch = { capMatch: string }
+export const isCapMatch = (match: Match): match is CapMatch =>
+  isMatcherType<CapMatch>(match, 'capMatch')
+
+export type PlusMatch = { plusMatch: string }
+export const isPlusMatch = (match: Match): match is PlusMatch =>
+  isMatcherType<PlusMatch>(match, 'plusMatch')
+
+export type MinusMatch = { minusMatch: string }
+export type Match =
+  | CoreNotationMatch
+  | DropHighMatch
+  | DropLowMatch
+  | DropConstraintsMatch
+  | ExplodeMatch
+  | UniqueMatch
+  | ReplaceMatch
+  | RerollMatch
+  | CapMatch
+  | PlusMatch
+  | MinusMatch
 
 function parseCoreNotationCustomSides(
   sides: string
@@ -68,13 +110,13 @@ function parseCapNotation({
   capString.forEach((note) => {
     capParameters = note.includes('<')
       ? {
-          ...capParameters,
-          lessThan: Number(note.replace(/</g, ''))
-        }
+        ...capParameters,
+        lessThan: Number(note.replace(/</g, ''))
+      }
       : {
-          ...capParameters,
-          greaterThan: Number(note.replace(/>/g, ''))
-        }
+        ...capParameters,
+        greaterThan: Number(note.replace(/>/g, ''))
+      }
   })
   return { cap: capParameters }
 }
